// src/types/datasources/postgres.types.ts
export enum PostgreSQLColumnType {
	INT = 'INT',
	DOUBLE = 'double precision',
	NUMERIC = 'numeric',
	REAL = 'real',
	TIMESTAMP = 'TIMESTAMP',
	CHAR = 'character',
	VARCHAR = 'VARCHAR',
	TEXT = 'text',
	DATE = 'date',
	DATETIME = 'timestamp',
	TIME = 'time',
	YEAR = 'integer',
	BOOLEAN = 'boolean',
	ENUM = 'enum',
	BINARY = 'bytea',
	JSON = 'json',
}